my-app/mlapi/requirements.txt
absl-py==0.11.0    
appdirs==1.4.3     
astunparse==1.6.3     
CacheControl==0.12.6    
cachetools==4.1.1     
certifi==2019.11.28
chardet==3.0.4     
click==7.1.2     
colorama==0.4.3     
contextlib2==0.6.0     
distlib==0.3.0     
distro==1.4.0     
Flask==1.1.2     
Flask-Cors==3.0.9     
gast==0.3.3     
google-auth==1.23.0    
google-auth-oauthlib==0.4.2     
google-pasta==0.2.0     
grpcio==1.33.2    
gunicorn==20.0.4 
h5py==2.10.0    
html5lib==1.0.1     
idna==2.8       
ipaddr==2.2.0     
itsdangerous==1.1.0     
Jinja2==2.11.2    
joblib==0.17.0    
Keras==2.4.3     
Keras-Preprocessing==1.1.2     
lockfile==0.12.2    
Markdown==3.3.3     
MarkupSafe==1.1.1     
msgpack==0.6.2     
numpy==1.18.5    
oauthlib==3.1.0     
opt-einsum==3.3.0     
packaging==20.3      
pandas==1.1.4     
pep517==0.8.2     
pip==20.0.2       
progress==1.5       
protobuf==3.14.0    
pyasn1==0.4.8     
pyasn1-modules==0.2.8     
pyparsing==2.4.6     
python-dateutil==2.8.1     
python-dotenv==0.15.0    
pytoml==0.1.21    
pytz==2020.4    
PyYAML==5.3.1     
requests==2.22.0    
requests-oauthlib==1.3.0     
retrying==1.3.3     
rsa==4.6       
scikit-learn==0.23.2    
scipy==1.5.4     
setuptools==44.0.0    
six==1.14.0    
sklearn==0.0       
tensorboard==2.4.0     
tensorboard-plugin-wit==1.7.0     
tensorflow-cpu==2.3.1     
tensorflow-estimator==2.3.0     
termcolor==1.1.0     
threadpoolctl==2.1.0     
urllib3==1.25.8    
webencodings==0.5.1     
Werkzeug==1.0.1     
wheel==0.34.2    
wrapt==1.12.1   
